Add files via upload
What I found most interesting about this Lab is how easy it is to manipulate formulas in order to receive the information you are looking for. Python is very helpful when you know the code to get the information you would require although that takes some time to learn and master. The most useful thing I found during this lab was the step by step coding practice.